requirements-dev-minimal.txt
bandit
behave-django
black
django-extensions
django-stubs
ipython
isort
mypy<0.780,>=0.770
pycodestyle
pygraphviz
pylint-django
pylint>=2.5.0
pytest
pytest-cov
pytest-django
requirements-tools
tox
factory_boy
lxml